[ONNX] Replace optional parameters of Resize with placeholder for ops13. (#50574)
* Replace optional parameters of Resize with placeholder for ops13.

* Use common methods to handle different versions.

* Correct flake8 issue.

* Update per comments.

* Add something to trigger CI again.

* Trigger another round of CI.

[ghstack-poisoned]
